+++ src/main/java/com/capgi/csvbuilder/ICSVBuilder.java
@@ -1,10 +1,10 @@
-package com.capgi;
+package com.capgi.csvbuilder;
 
 import java.io.Reader;
 import java.util.Iterator;
 
 public interface ICSVBuilder<E> {
-	public Iterator<E> getCsvBeanIterator(Reader reader, Class csvClass) throws CustomCensusAnalyserException;
+	public Iterator<E> getCsvBeanIterator(Reader reader, Class csvClass) throws CsvException;
 
 	public <E> int getNoOfEntries(Iterator<E> iterator);
 }
