+++ src/main/java/com/capgi/csvbuilder/OpenCsvBuilder.java
@@ -1,20 +1,19 @@
-package com.capgi;
+package com.capgi.csvbuilder;
 
 import java.io.Reader;
 import java.util.Iterator;
 
-import com.capgi.CustomCensusAnalyserException.ExceptionType;
 import com.opencsv.bean.CsvToBeanBuilder;
 
 public class OpenCsvBuilder<E> implements ICSVBuilder<E> {
-	public Iterator<E> getCsvBeanIterator(Reader reader, Class csvClass) throws CustomCensusAnalyserException {
+	public Iterator<E> getCsvBeanIterator(Reader reader, Class csvClass) throws CsvException {
 		try {
 			CsvToBeanBuilder<E> csvToBean = new CsvToBeanBuilder<E>(reader);
 			csvToBean.withType(csvClass);
 			csvToBean.withIgnoreLeadingWhiteSpace(true);
 			return csvToBean.build().iterator();
 		} catch (IllegalStateException e) {
-			throw new CustomCensusAnalyserException("Invalid class", ExceptionType.InvalidClass);
+			throw new CsvException(e.getMessage(), CsvException.ExceptionType.UNABLE_TO_PARSE);
 		}
 
 	}
